+++ src/main/java/com/heroiclabs/nakama/DefaultMatchmakeRangeFilter.java
@@ -16,34 +16,16 @@
 
 package com.heroiclabs.nakama;
 
-import java.util.List;
+import lombok.*;
 
-/**
- * The object which represents a matchmake matched.
- */
-public interface MatchmakeMatched {
-    /**
-     * @return Matchmaking Ticket.
-     */
-    MatchmakeTicket getTicket();
+@Data
+@ToString(includeFieldNames = true)
+@RequiredArgsConstructor(access = AccessLevel.PACKAGE)
+class DefaultMatchmakeRangeFilter implements MatchmakeRangeFilter {
+    private final long lowerbound;
+    private final long upperbound;
 
-    /**
-     * @return List of presences in this match.
-     */
-    List<UserPresence> getPresence();
-
-    /**
-     * @return Match token.
-     */
-    MatchToken getToken();
-
-    /**
-     * @return Match presence for the current user.
-     */
-    UserPresence getSelf();
-
-    /**
-     * @return Match properties for matched users.
-     */
-    List<MatchmakeUserProperty> getProperties();
+    static MatchmakeRangeFilter fromProto(final @NonNull com.heroiclabs.nakama.Api.MatchmakeFilter.RangeFilter filter) {
+        return new DefaultMatchmakeRangeFilter(filter.getLowerBound(), filter.getUpperBound());
+    }
 }