+++ src/main/java/com/heroiclabs/nakama/DefaultMatchmakeTermFilter.java
@@ -16,34 +16,18 @@
 
 package com.heroiclabs.nakama;
 
+import lombok.*;
+
 import java.util.List;
 
-/**
- * The object which represents a matchmake matched.
- */
-public interface MatchmakeMatched {
-    /**
-     * @return Matchmaking Ticket.
-     */
-    MatchmakeTicket getTicket();
+@Data
+@ToString(includeFieldNames = true)
+@RequiredArgsConstructor(access = AccessLevel.PACKAGE)
+class DefaultMatchmakeTermFilter implements MatchmakeTermFilter {
+    private final List<String> terms;
+    private final boolean matchingAllTerms;
 
-    /**
-     * @return List of presences in this match.
-     */
-    List<UserPresence> getPresence();
-
-    /**
-     * @return Match token.
-     */
-    MatchToken getToken();
-
-    /**
-     * @return Match presence for the current user.
-     */
-    UserPresence getSelf();
-
-    /**
-     * @return Match properties for matched users.
-     */
-    List<MatchmakeUserProperty> getProperties();
+    static MatchmakeTermFilter fromProto(final @NonNull com.heroiclabs.nakama.Api.MatchmakeFilter.TermFilter filter) {
+        return new DefaultMatchmakeTermFilter(filter.getTermsList(), filter.getMatchAllTerms());
+    }
 }