+++ src/main/java/com/heroiclabs/nakama/MatchmakeCheckFilter.java
@@ -16,34 +16,9 @@
 
 package com.heroiclabs.nakama;
 
-import java.util.List;
-
-/**
- * The object which represents a matchmake matched.
- */
-public interface MatchmakeMatched {
+public interface MatchmakeCheckFilter extends MatchmakeFilter {
     /**
-     * @return Matchmaking Ticket.
+     * @return Filter value
      */
-    MatchmakeTicket getTicket();
-
-    /**
-     * @return List of presences in this match.
-     */
-    List<UserPresence> getPresence();
-
-    /**
-     * @return Match token.
-     */
-    MatchToken getToken();
-
-    /**
-     * @return Match presence for the current user.
-     */
-    UserPresence getSelf();
-
-    /**
-     * @return Match properties for matched users.
-     */
-    List<MatchmakeUserProperty> getProperties();
+    boolean isValue();
 }